+++ src/cWeb.ts
@@ -32,38 +32,26 @@ export default class {
                 case 'openT':
                     vscode.commands.executeCommand('cyezoi.openT', message.rid);
                     break;
+                case 'refresh':
+                    this.fetchData(tid);
+                    break;
             }
         });
 
-        new fetch({
-            path: `/d/${settings.domain}/contest/${tid}/scoreboard`
-            , addCookie: true
-        }).start().then(async (contestDetail) => {
-            if (contestDetail?.json !== undefined) {
-                const data = contestDetail.json;
-                data.tdoc.content = await marked(data.tdoc.content);
-                const message = {
-                    command: 'contest',
-                    data,
-                };
-                this._panel.webview.postMessage(message);
-            }
-        }).catch(async (e: Error) => {
-            io.error(e.message);
-        });
+        this.fetchData(tid);
     }
 
-    private getRealPath(relativePath: string[]): vscode.Uri {
+    private getRealPath = (relativePath: string[]): vscode.Uri => {
         return this._panel?.webview.asWebviewUri(
             vscode.Uri.file(path.join(this._extensionPath, ...relativePath)),
         );
-    }
+    };
 
-    private getHtml() {
+    private getHtml = () => {
         const staticFiles = [
             { 'path': ['res', 'libs', 'vscode-elements', 'bundled.js'], attributes: { 'type': 'module' } },
             { 'path': ['res', 'libs', 'codicon', 'codicon.css'], attributes: { 'id': 'vscode-codicon-stylesheet' } },
-            { 'path': ['res', 'html', 'static.js'] },
+            { 'path': ['src', 'utils.js'] },
             { 'path': ['res', 'html', 'contest.js'] },
         ];
         let recordHtml = path.join(this._extensionPath, 'res', 'html', 'contest.html');
@@ -90,5 +78,24 @@ export default class {
         }).join('\n'));
         htmlContent = htmlContent.replace(/{{hydroIconsFile}}/g, this.getRealPath(['res', 'fonts', 'hydro-icons.woff2']));
         return htmlContent;
-    }
+    };
+
+    private fetchData = async (tid: string) => {
+        new fetch({
+            path: `/d/${settings.domain}/contest/${tid}/scoreboard`
+            , addCookie: true
+        }).start().then(async (contestDetail) => {
+            if (contestDetail?.json !== undefined) {
+                const data = contestDetail.json;
+                data.tdoc.content = await marked(data.tdoc.content);
+                const message = {
+                    command: 'contest',
+                    data,
+                };
+                this._panel.webview.postMessage(message);
+            }
+        }).catch(async (e: Error) => {
+            io.error(e.message);
+        });
+    };
 }