+++ src/pTree.ts
@@ -1,12 +1,11 @@
 import * as vscode from 'vscode';
+import * as utils from './utils';
 import fetch from './fetch';
 import path from 'path';
-import { statusIcon, statusName, ProblemDoc, ProblemStatusDoc, RecordDoc, languageDisplayName } from './static';
 import { io, outputChannel } from './io';
 import settings from './settings';
 import { Record } from './rTree';
 import storage from './storage';
-import { toMemory, toRelativeTime, toTime } from './utils';
 
 export default class implements vscode.TreeDataProvider<Problem> {
     private _onDidChangeTreeData: vscode.EventEmitter<Problem | undefined> = new vscode.EventEmitter<any | undefined>();
@@ -68,7 +67,7 @@ export default class implements vscode.TreeDataProvider<Problem> {
 }
 
 export class Problem extends vscode.TreeItem {
-    constructor(pdoc: ProblemDoc, psdoc: ProblemStatusDoc) {
+    constructor(pdoc: utils.ProblemDoc, psdoc: utils.ProblemStatusDoc) {
         super('P' + pdoc.docId, (psdoc ? vscode.TreeItemCollapsibleState.Collapsed : vscode.TreeItemCollapsibleState.None));
         this.contextValue = 'problem';
         this.description = pdoc.title;
@@ -77,8 +76,8 @@ export class Problem extends vscode.TreeItem {
             tooltipDoc.appendMarkdown(`- **Tags**: ${pdoc.tag.join(', ')}\n`);
         }
         if (psdoc) {
-            this.iconPath = path.join(__dirname, '..', 'res', 'icons', statusIcon[psdoc.status] + '.svg');
-            tooltipDoc.appendMarkdown(`- **Status**: ${statusName[psdoc.status]}\n`);
+            this.iconPath = path.join(__dirname, '..', 'res', 'icons', utils.statusIcon[psdoc.status] + '.svg');
+            tooltipDoc.appendMarkdown(`- **Status**: ${utils.statusName[psdoc.status]}\n`);
             tooltipDoc.appendMarkdown(`- **Score**: ${psdoc.score}\n`);
         }
         tooltipDoc.appendMarkdown(`- **Difficulty**: ${pdoc.difficulty}\n`);
@@ -93,17 +92,17 @@ export class Problem extends vscode.TreeItem {
 }
 
 export class ProblemRecord extends vscode.TreeItem {
-    constructor(rdoc: RecordDoc) {
-        super(rdoc.score + ' ' + statusName[rdoc.status], vscode.TreeItemCollapsibleState.None);
+    constructor(rdoc: utils.RecordDoc) {
+        super(rdoc.score + ' ' + utils.statusName[rdoc.status], vscode.TreeItemCollapsibleState.None);
         this.contextValue = 'record';
         const tooltipDoc = new vscode.MarkdownString();
-        this.iconPath = path.join(__dirname, '..', 'res', 'icons', statusIcon[rdoc.status] + '.svg');
-        tooltipDoc.appendMarkdown(`- **Status**: ${statusName[rdoc.status]}\n`);
+        this.iconPath = path.join(__dirname, '..', 'res', 'icons', utils.statusIcon[rdoc.status] + '.svg');
+        tooltipDoc.appendMarkdown(`- **Status**: ${utils.statusName[rdoc.status]}\n`);
         tooltipDoc.appendMarkdown(`- **Score**: ${rdoc.score}\n`);
-        if (rdoc.time) { tooltipDoc.appendMarkdown(`- **Time**: ${toTime(rdoc.time)}\n`); }
-        if (rdoc.memory) { tooltipDoc.appendMarkdown(`- **Memory**: ${toMemory(rdoc.memory * 1024)}\n`); }
-        tooltipDoc.appendMarkdown(`- **Lang**: ${languageDisplayName[rdoc.lang]}\n`);
-        tooltipDoc.appendMarkdown(`- **Judge At**: ${toRelativeTime(new Date(rdoc.judgeAt).getTime())}\n`);
+        if (rdoc.time) { tooltipDoc.appendMarkdown(`- **Time**: ${utils.toTime(rdoc.time)}\n`); }
+        if (rdoc.memory) { tooltipDoc.appendMarkdown(`- **Memory**: ${utils.toMemory(rdoc.memory * 1024)}\n`); }
+        tooltipDoc.appendMarkdown(`- **Lang**: ${utils.languageDisplayName[rdoc.lang]}\n`);
+        tooltipDoc.appendMarkdown(`- **Judge At**: ${utils.toRelativeTime(new Date(rdoc.judgeAt).getTime())}\n`);
         this.tooltip = tooltipDoc;
         this.command = {
             command: 'cyezoi.openT',