+++ src/pWeb.ts
@@ -29,9 +29,69 @@ export default class {
                 case 'submitProblem':
                     vscode.commands.executeCommand('cyezoi.submitProblem', message.problemId, tid);
                     break;
+                case 'refresh':
+                    this.fetchData(pid, tid);
+                    break;
             }
         });
 
+        this.fetchData(pid, tid);
+    }
+
+    private getRealPath = (relativePath: string[]): vscode.Uri => {
+        return this._panel?.webview.asWebviewUri(
+            vscode.Uri.file(path.join(this._extensionPath, ...relativePath)),
+        );
+    };
+
+    private getHtml = () => {
+        const staticFiles = [
+            { 'path': ['res', 'libs', 'MathJax-3.2.2', 'tex-mml-chtml.js'], attributes: { 'async': undefined, 'id': 'MathJax-script' } },
+            { 'path': ['res', 'libs', 'vscode-elements', 'bundled.js'], attributes: { 'type': 'module' } },
+            { 'path': ['res', 'libs', 'codicon', 'codicon.css'], attributes: { 'id': 'vscode-codicon-stylesheet' } },
+            { 'path': ['res', 'libs', 'codemirror', 'codemirror.min.js'] },
+            { 'path': ['res', 'libs', 'codemirror', 'codemirror.min.css'] },
+            { 'path': ['res', 'libs', 'codemirror', 'theme', 'material.min.css'] },
+            { 'path': ['res', 'libs', 'codemirror', 'mode', 'clike', 'clike.min.js'] },
+            { 'path': ['res', 'libs', 'codemirror', 'addon', 'fold', 'foldcode.min.js'] },
+            { 'path': ['res', 'libs', 'codemirror', 'addon', 'fold', 'foldgutter.min.css'] },
+            { 'path': ['res', 'libs', 'codemirror', 'addon', 'fold', 'foldgutter.min.js'] },
+            { 'path': ['res', 'libs', 'codemirror', 'addon', 'fold', 'brace-fold.min.js'] },
+            { 'path': ['res', 'libs', 'codemirror', 'addon', 'fold', 'comment-fold.min.js'] },
+            { 'path': ['res', 'libs', 'codemirror', 'addon', 'fold', 'indent-fold.min.js'] },
+            { 'path': ['res', 'libs', 'codemirror', 'addon', 'section', 'active-line.min.js'] },
+            { 'path': ['res', 'libs', 'codemirror', 'addon', 'display', 'autorefresh.min.js'] },
+            { 'path': ['src', 'utils.js'] },
+            { 'path': ['res', 'html', 'problem.css'] },
+            { 'path': ['res', 'html', 'problem.js'] },
+        ];
+        let recordHtml = path.join(this._extensionPath, 'res', 'html', 'problem.html');
+        let htmlContent = require('fs').readFileSync(recordHtml, 'utf8');
+        htmlContent = htmlContent.replace(/{{staticFiles}}/g, staticFiles.map(file => {
+            let attributes = '';
+            if (file.attributes !== undefined) {
+                attributes = ' ' + Object.entries(file.attributes).map(([key, value]) => {
+                    if (value === undefined) {
+                        return key;
+                    }
+                    return `${key}="${value}"`;
+                }).join(' ');
+            }
+            if (file.path[file.path.length - 1].endsWith('.css')) {
+                return `<link rel="stylesheet" type="text/css" href="${this.getRealPath(file.path)}"${attributes}>`;
+            }
+            else if (file.path[file.path.length - 1].endsWith('.js')) {
+                return `<script src="${this.getRealPath(file.path)}"${attributes}></script>`;
+            }
+            else {
+                throw new Error('Unknown file type');
+            }
+        }).join('\n'));
+        htmlContent = htmlContent.replace(/{{hydroIconsFile}}/g, this.getRealPath(['res', 'fonts', 'hydro-icons.woff2']));
+        return htmlContent;
+    };
+
+    private fetchData = (pid: string, tid?: string) => {
         new fetch({
             path: `/d/${settings.domain}/p/${pid}` + (tid !== undefined ? `?tid=${tid}` : '')
             , addCookie: true
@@ -77,58 +137,5 @@ export default class {
                 io.error(e.message);
             });
         }
-    }
-
-    private getRealPath(relativePath: string[]): vscode.Uri {
-        return this._panel?.webview.asWebviewUri(
-            vscode.Uri.file(path.join(this._extensionPath, ...relativePath)),
-        );
-    }
-
-    private getHtml() {
-        const staticFiles = [
-            { 'path': ['res', 'libs', 'MathJax-3.2.2', 'tex-mml-chtml.js'], attributes: { 'async': undefined, 'id': 'MathJax-script' } },
-            { 'path': ['res', 'libs', 'vscode-elements', 'bundled.js'], attributes: { 'type': 'module' } },
-            { 'path': ['res', 'libs', 'codicon', 'codicon.css'], attributes: { 'id': 'vscode-codicon-stylesheet' } },
-            { 'path': ['res', 'libs', 'codemirror', 'codemirror.min.js'] },
-            { 'path': ['res', 'libs', 'codemirror', 'codemirror.min.css'] },
-            { 'path': ['res', 'libs', 'codemirror', 'theme', 'material.min.css'] },
-            { 'path': ['res', 'libs', 'codemirror', 'mode', 'clike', 'clike.min.js'] },
-            { 'path': ['res', 'libs', 'codemirror', 'addon', 'fold', 'foldcode.min.js'] },
-            { 'path': ['res', 'libs', 'codemirror', 'addon', 'fold', 'foldgutter.min.css'] },
-            { 'path': ['res', 'libs', 'codemirror', 'addon', 'fold', 'foldgutter.min.js'] },
-            { 'path': ['res', 'libs', 'codemirror', 'addon', 'fold', 'brace-fold.min.js'] },
-            { 'path': ['res', 'libs', 'codemirror', 'addon', 'fold', 'comment-fold.min.js'] },
-            { 'path': ['res', 'libs', 'codemirror', 'addon', 'fold', 'indent-fold.min.js'] },
-            { 'path': ['res', 'libs', 'codemirror', 'addon', 'section', 'active-line.min.js'] },
-            { 'path': ['res', 'libs', 'codemirror', 'addon', 'display', 'autorefresh.min.js'] },
-            { 'path': ['res', 'html', 'static.js'] },
-            { 'path': ['res', 'html', 'problem.css'] },
-            { 'path': ['res', 'html', 'problem.js'] },
-        ];
-        let recordHtml = path.join(this._extensionPath, 'res', 'html', 'problem.html');
-        let htmlContent = require('fs').readFileSync(recordHtml, 'utf8');
-        htmlContent = htmlContent.replace(/{{staticFiles}}/g, staticFiles.map(file => {
-            let attributes = '';
-            if (file.attributes !== undefined) {
-                attributes = ' ' + Object.entries(file.attributes).map(([key, value]) => {
-                    if (value === undefined) {
-                        return key;
-                    }
-                    return `${key}="${value}"`;
-                }).join(' ');
-            }
-            if (file.path[file.path.length - 1].endsWith('.css')) {
-                return `<link rel="stylesheet" type="text/css" href="${this.getRealPath(file.path)}"${attributes}>`;
-            }
-            else if (file.path[file.path.length - 1].endsWith('.js')) {
-                return `<script src="${this.getRealPath(file.path)}"${attributes}></script>`;
-            }
-            else {
-                throw new Error('Unknown file type');
-            }
-        }).join('\n'));
-        htmlContent = htmlContent.replace(/{{hydroIconsFile}}/g, this.getRealPath(['res', 'fonts', 'hydro-icons.woff2']));
-        return htmlContent;
-    }
+    };
 }