+++ src/rTree.ts
@@ -1,10 +1,9 @@
 import * as vscode from 'vscode';
+import * as utils from './utils';
 import fetch from './fetch';
 import path from 'path';
-import { languageDisplayName, ProblemDoc, RecordDoc, statusIcon, statusName, UserDoc } from './static';
 import { io, outputChannel } from './io';
 import settings from './settings';
-import { toMemory, toRelativeTime, toTime } from './utils';
 
 export default class implements vscode.TreeDataProvider<Record> {
     private _onDidChangeTreeData: vscode.EventEmitter<Record | undefined> = new vscode.EventEmitter<any | undefined>();
@@ -51,19 +50,19 @@ export default class implements vscode.TreeDataProvider<Record> {
 }
 
 export class Record extends vscode.TreeItem {
-    constructor(rdoc: RecordDoc, pdoc: ProblemDoc, udoc: UserDoc) {
-        super(rdoc.score + ' ' + statusName[rdoc.status], vscode.TreeItemCollapsibleState.None);
+    constructor(rdoc: utils.RecordDoc, pdoc: utils.ProblemDoc, udoc: utils.UserDoc) {
+        super(rdoc.score + ' ' + utils.statusName[rdoc.status], vscode.TreeItemCollapsibleState.None);
         this.contextValue = 'record';
         this.description = 'P' + rdoc.pid + ' ' + pdoc.title;
-        this.iconPath = path.join(__dirname, '..', 'res', 'icons', statusIcon[rdoc.status] + '.svg');
+        this.iconPath = path.join(__dirname, '..', 'res', 'icons', utils.statusIcon[rdoc.status] + '.svg');
         const TooltipDoc = new vscode.MarkdownString();
-        TooltipDoc.appendMarkdown(`- **Status**: ${statusName[rdoc.status]}\n`);
+        TooltipDoc.appendMarkdown(`- **Status**: ${utils.statusName[rdoc.status]}\n`);
         TooltipDoc.appendMarkdown(`- **User**: ${udoc.uname}\n`);
         TooltipDoc.appendMarkdown(`- **Score**: ${rdoc.score}\n`);
-        if (rdoc.time) { TooltipDoc.appendMarkdown(`- **Time**: ${toTime(rdoc.time)}\n`); }
-        if (rdoc.memory) { TooltipDoc.appendMarkdown(`- **Memory**: ${toMemory(rdoc.memory)}\n`); }
-        TooltipDoc.appendMarkdown(`- **Lang**: ${languageDisplayName[rdoc.lang]}\n`);
-        TooltipDoc.appendMarkdown(`- **Judge At**: ${toRelativeTime(new Date(rdoc.judgeAt).getTime())}\n`);
+        if (rdoc.time) { TooltipDoc.appendMarkdown(`- **Time**: ${utils.toTime(rdoc.time)}\n`); }
+        if (rdoc.memory) { TooltipDoc.appendMarkdown(`- **Memory**: ${utils.toMemory(rdoc.memory)}\n`); }
+        TooltipDoc.appendMarkdown(`- **Lang**: ${utils.languageDisplayName[rdoc.lang]}\n`);
+        TooltipDoc.appendMarkdown(`- **Judge At**: ${utils.toRelativeTime(new Date(rdoc.judgeAt).getTime())}\n`);
         this.tooltip = TooltipDoc;
         this.command = {
             command: 'cyezoi.openT',