+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to the "cyezoi" extension will be documented in this file.
 ## v0.0.17
 
 - **Add keybindings to submit and open problem**
+- **Add refresh button to webview**
 - Fix the issue of displaying the relative time
 - Enhance login experience